pyproject.toml
[project]
name = "mcp-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.8",
    "fastmcp>=2.11.2",
    "langfuse>=3.2.3",
    "llama-index>=0.13.1",
    "llama-index-llms-google-genai>=0.3.0",
    "loguru>=0.7.3",
    "openinference-instrumentation-llama-index>=4.3.4",
]

[tool.ruff]
exclude = ["build"]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
    "W",   # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true

[tool.bumpversion]
current_version = "0.1.0"
tag = true
commit = true
pre_commit_hooks = ["uv lock", "git add uv.lock"]

[dependency-groups]
dev = []
